pyproject.toml
[project]
name = "spec-clarify-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "numpy>=2.3.4",
    "openai>=1.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "requests>=2.31.0",
]
